feat: adding logic so that the Sample, Observer objects and the new Activity fields are set

+++ projeto-tcc-api/src/main/java/com/example/projeto_tcc/repository/ObserverRepository.java
@@ -3,5 +3,5 @@ package com.example.projeto_tcc.repository;
 import com.example.projeto_tcc.entity.Observer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObserverRepository extends JpaRepository<Observer, Integer> {
+public interface ObserverRepository extends JpaRepository<Observer, Long> {
 }
